Secure Telegram Mini App initData authentication

+++ backend/app/schemas/wallet.py
@@ -1,6 +1,8 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class AddEFC(BaseModel):
     telegram_id: int
-    amount: float
+    amount: Decimal
